+++ src/vs/workbench/contrib/notebook/browser/notebook.contribution.ts
@@ -926,13 +926,11 @@ configurationRegistry.registerConfiguration({
 			additionalProperties: {
 				type: 'boolean'
 			},
-			tags: ['notebookLayout'],
 			default: {}
 		},
 		[NotebookSetting.formatOnCellExecution]: {
-			markdownDescription: nls.localize('notebook.formatOnCellExecution', "Format a notebook cell upon execution. A formatter must be available, and the editor must not be shutting down."),
+			markdownDescription: nls.localize('notebook.formatOnCellExecution', "Format a notebook cell upon execution. A formatter must be available."),
 			type: 'boolean',
-			tags: ['notebookLayout'],
 			default: false
 		},
 		[NotebookSetting.confirmDeleteRunningCell]: {
