+++ src/vs/workbench/contrib/notebook/common/notebookExecutionService.ts
@@ -3,10 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { IDisposable } from 'vs/base/common/lifecycle';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { NotebookCellTextModel } from 'vs/workbench/contrib/notebook/common/model/notebookCellTextModel';
 import { INotebookTextModel, IOutputDto, IOutputItemDto } from 'vs/workbench/contrib/notebook/common/notebookCommon';
+import { INotebookCellExecution } from 'vs/workbench/contrib/notebook/common/notebookExecutionStateService';
 
 export enum CellExecutionUpdateType {
 	Output = 1,
@@ -36,4 +38,10 @@ export interface INotebookExecutionService {
 	executeNotebookCells(notebook: INotebookTextModel, cells: Iterable<NotebookCellTextModel>, contextKeyService: IContextKeyService): Promise<void>;
 	cancelNotebookCells(notebook: INotebookTextModel, cells: Iterable<NotebookCellTextModel>): Promise<void>;
 	cancelNotebookCellHandles(notebook: INotebookTextModel, cells: Iterable<number>): Promise<void>;
+	registerExecutionParticipant(participant: ICellExecutionParticipant): IDisposable;
+	runExecutionParticipants(executions: INotebookCellExecution[]): Promise<void>;
+}
+
+export interface ICellExecutionParticipant {
+	onWillExecuteCell(executions: INotebookCellExecution[]): Promise<void>;
 }
